fix: added an abstract interface class for entities
