+++ modelon/impact/client/__init__.py
@@ -1,6 +1,8 @@
 from modelon.impact.client.client import Client
-from modelon.impact.client.experiment_definition.base import (
+from modelon.impact.client.experiment_definition.fmu_based import (
     SimpleFMUExperimentDefinition,
+)
+from modelon.impact.client.experiment_definition.model_based import (
     SimpleModelicaExperimentDefinition,
 )
 from modelon.impact.client.experiment_definition.extension import (
