+++ modelon/impact/client/entities/interfaces/case.py
@@ -1,0 +1,9 @@
+from abc import ABC, abstractmethod
+
+
+class CaseInterface(ABC):
+    @property
+    @abstractmethod
+    def id(self) -> str:
+        """Case id."""
+        raise NotImplementedError
