+++ modelon/impact/client/entities/interfaces/custom_function.py
@@ -1,0 +1,9 @@
+from abc import ABC, abstractmethod
+
+
+class CustomFunctionInterface(ABC):
+    @property
+    @abstractmethod
+    def name(self) -> str:
+        "Custom function name"
+        raise NotImplementedError
