+++ modelon/impact/client/entities/interfaces/external_result.py
@@ -1,0 +1,9 @@
+from abc import ABC, abstractmethod
+
+
+class ExternalResultInterface(ABC):
+    @property
+    @abstractmethod
+    def id(self) -> str:
+        """Result id."""
+        raise NotImplementedError
