+++ modelon/impact/client/entities/interfaces/model.py
@@ -1,0 +1,9 @@
+from abc import ABC, abstractmethod
+
+
+class ModelInterface(ABC):
+    @property
+    @abstractmethod
+    def name(self) -> str:
+        """Class name."""
+        raise NotImplementedError
