+++ modelon/impact/client/entities/interfaces/model_executable.py
@@ -1,0 +1,9 @@
+from abc import ABC, abstractmethod
+
+
+class ModelExecutableInterface(ABC):
+    @property
+    @abstractmethod
+    def id(self) -> str:
+        """FMU id."""
+        raise NotImplementedError
