+++ modelon/impact/client/entities/interfaces/workspace.py
@@ -1,0 +1,9 @@
+from abc import ABC, abstractmethod
+
+
+class WorkspaceInterface(ABC):
+    @property
+    @abstractmethod
+    def id(self) -> str:
+        """Workspace id."""
+        raise NotImplementedError
