+++ modelon/impact/client/entities/model.py
@@ -2,11 +2,14 @@ from __future__ import annotations
 import logging
 from typing import Any, Dict, Optional, Union, TYPE_CHECKING
 
+from modelon.impact.client.entities.interfaces.model import ModelInterface
 from modelon.impact.client.operations.model_executable import (
     ModelExecutableOperation,
     CachedModelExecutableOperation,
 )
-from modelon.impact.client.experiment_definition import base
+from modelon.impact.client.experiment_definition.model_based import (
+    SimpleModelicaExperimentDefinition,
+)
 from modelon.impact.client.options import (
     ProjectExecutionOptions,
     CompilerOptions,
@@ -55,7 +58,7 @@ def _assert_valid_compilation_options(
         )
 
 
-class Model:
+class Model(ModelInterface):
     """Class containing Model functionalities."""
 
     def __init__(
@@ -201,7 +204,7 @@ class Model:
         simulation_options: Optional[SimulationOptionsOrDict] = None,
         simulation_log_level: str = "WARNING",
         initialize_from: Optional[CaseOrExperimentOrExternalResult] = None,
-    ) -> base.SimpleModelicaExperimentDefinition:
+    ) -> SimpleModelicaExperimentDefinition:
         """Returns a new experiment definition using this Model.
 
         Args:
@@ -259,7 +262,7 @@ class Model:
             self._project_id, self._workspace_id, custom_function=custom_function.name
         )
         project_options = ProjectExecutionOptions(options, custom_function.name)
-        return base.SimpleModelicaExperimentDefinition(
+        return SimpleModelicaExperimentDefinition(
             model=self,
             custom_function=custom_function,
             compiler_options=compiler_options or project_options.compiler_options,
