+++ modelon/impact/client/experiment_definition/interfaces/extension.py
@@ -1,0 +1,10 @@
+from __future__ import annotations
+import logging
+
+logger = logging.getLogger(__name__)
+
+
+class BaseExperimentExtension:
+    """Base class for an experiment extension class."""
+
+    pass
