+++ tests/impact/client/sal/test_workspace_service.py
@@ -1,4 +1,3 @@
-import unittest.mock as mock
 from modelon.impact.client.sal.uri import URI
 import modelon.impact.client.sal.service
 from tests.impact.client.helpers import (
